requirements.txt
steam[client]==1.4.4
beautifulsoup4==4.13.3
httpx==0.25.0
python-dotenv==1.1.0
eventemitter
aiohttp
steampy==1.2.0